GoIoTWorkInterruptTrackerLambda.go: fix handling of input JSON per event structure to grab from published event field.
To proplerly handle button event data, there have been structs created to unmarshal JSON and create structs to grab the deviceEvent and deviceInfo JSON segments. As a result, all Placement info could be ignored and everything passed in can be handled through JSON variables. The following structure is obtained:
1. struct containing two structs: deviceInfo and deviceEvent.
2. deviceEvent is a struct consisting of another struct: iotButtonClicked
3. iotButtonClicked contains the event details: click type and reported time
4. iotButtonDeviceInfo contains the serial number and the remaining battery charge

By taking the above steps, there is no need for manually setting attributes or reading those attributes. Instead, all fields can be extracted and pushed into ddb. The advantage of capturing click type and timestamp is the deltas on a particular type can be used to determine duration OR the single -> double click could signal start and stop timer to calculate more accurate time tracking. This data can now be post-processed AND multiple devices can be connected to a single account in a single ddb table to allow large deployments, if desired.